+++ sdk/resourcemanager/datafactory/armdatafactory/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 1.1.0 (2022-05-30)
+### Features Added
+
+- New function `GlobalParameterResource.MarshalJSON() ([]byte, error)`
+- New struct `GlobalParameterListResponse`
+- New struct `GlobalParameterResource`
+- New struct `GlobalParametersClientCreateOrUpdateOptions`
+- New struct `GlobalParametersClientCreateOrUpdateResponse`
+- New struct `GlobalParametersClientDeleteOptions`
+- New struct `GlobalParametersClientDeleteResponse`
+- New struct `GlobalParametersClientGetOptions`
+- New struct `GlobalParametersClientGetResponse`
+- New struct `GlobalParametersClientListByFactoryOptions`
+- New struct `GlobalParametersClientListByFactoryResponse`
+
+
 ## 1.0.0 (2022-05-17)
 ### Breaking Changes